pyproject.toml
[tool.poetry]
name = "py-igs"
version = "0.5.0"
description = ""
authors = ["Enzo Albornoz <enzocoelhoalbornoz@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
pycairo = "^1.20.0"
PyGObject = "^3.40.1"
wgpu = "^0.5.4"


[tool.poetry.dev-dependencies]
yapf = "^0.31.0"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
